requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.3
mysql-connector-python==8.0.28
numpy==1.22.2
pandas==1.4.1
protobuf==3.19.4
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
six==1.16.0
urllib3==1.26.8
